Cleanup code + more tests
