+++ rewrite-java-tck/src/main/java/org/openrewrite/java/tree/TypeUtilsTest.java
@@ -15,10 +15,6 @@
  */
 package org.openrewrite.java.tree;
 
-import org.assertj.core.api.BooleanAssert;
-import org.assertj.core.api.ObjectAssert;
-import org.assertj.core.api.SoftAssertions;
-import org.assertj.core.api.StringAssert;
 import org.junit.jupiter.api.Test;
 import org.openrewrite.ExecutionContext;
 import org.openrewrite.InMemoryExecutionContext;
@@ -38,97 +34,6 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.openrewrite.java.Assertions.java;
 import static org.openrewrite.test.RewriteTest.toRecipe;
-
-//abstract class Base implements Comparable<Base> {}
-//abstract class Extends extends Base {}
-//
-//class A {
-//    static <T> void a(Map<T, T> mapA) {}
-//    static <T> void b(Map<T, ? extends T> mapB) {}
-//    static <T> void c(Map<T, ? super T> mapC) {}
-//
-//    static <T> void d(Map<? extends T, ? extends T> mapD) {}
-//    static <T> void e(Map<? extends T, ? super T> mapE) {}
-//    static <T> void f(Map<? super T, ? super T> mapF) {}
-//
-//    static <N extends Number> void an(Map<N, N> mapAn) {}
-//    static <N extends Number> void bn(Map<N, ? extends N> mapBn) {}
-//    static <N extends Number> void cn(Map<N, ? super N> mapCn) {}
-//
-//    static <N extends Number> void dn(Map<? extends N, ? extends N> mapDn) {}
-//    static <N extends Number> void en(Map<? extends N, ? super N> mapEn) {}
-//    static <N extends Number> void fn(Map<? super N, ? super N> mapFn) {}
-//
-//    void concrete(Map<Long, Long> mll, Map<Long, Number> mln, Map<Number, Long> mnl, Map<Number, Number> mnn, Map<? extends Number, ? extends Number> mex, Map<? super Number, ? super Number> msu) {
-//        a(mll);  an(mll);
-//        a(mln);  an(mln); // reason: Incompatible equality constraint: Long and Number
-//        a(mnl);  an(mnl); // reason: Incompatible equality constraint: Number and Long
-//        a(mnn);  an(mnn);
-//        a(mex);  an(mex); // reason: Incompatible equality constraint: capture of ? extends Number and capture of ? extends Number
-//        a(msu);  an(msu); // reason: Incompatible equality constraint: capture of ? extends Number and capture of ? extends Number
-//
-//        b(mll);  bn(mll);
-//        b(mln);  bn(mln); // reason: no instance(s) of type variable(s) exist so that Number conforms to Long
-//                          //         inference variable T has incompatible bounds:
-//                          //         equality constraints: Long
-//                          //         lower bounds: Number
-//        b(mnl);  bn(mnl);
-//        b(mnn);  bn(mnn);
-//        b(mex);  bn(mex); // reason: no instance(s) of type variable(s) exist so that capture of ? extends Number conforms to capture of ? extends Number
-//                          //         inference variable T has incompatible bounds:
-//                          //         equality constraints: capture of ? extends Number
-//                          //         lower bounds: lower bounds: capture of ? extends Number
-//
-//        c(mll);  cn(mll);
-//        c(mln);  cn(mln);
-//        c(mnl);  cn(mnl); // reason: no instance(s) of type variable(s) exist so that Number conforms to Long
-//        c(mnn);  cn(mnn);
-//        c(mex);  cn(mex); // reason: no instance(s) of type variable(s) exist so that capture of ? extends Number conforms to capture of ? extends Number
-//
-//        d(mll);  dn(mll);
-//        d(mln);  dn(mln);
-//        d(mnl);  dn(mnl);
-//        d(mnn);  dn(mnn);
-//        d(mex);  dn(mex);
-//
-//        e(mll);  en(mll);
-//        e(mln);  en(mln);
-//        e(mnl);  en(mnl); // reason: no instance(s) of type variable(s) exist so that Number conforms to Long
-//        e(mnn);  en(mnn);
-//        e(mex);  en(mex); // reason: no instance(s) of type variable(s) exist so that capture of ? extends Number conforms to capture of ? extends Number
-//
-//        f(mll);  fn(mll);
-//        f(mln);  fn(mln);
-//        f(mnl);  fn(mnl);
-//        f(mnn);  fn(mnn);
-//        f(mex);  fn(mex);
-//
-//
-//    }
-//    <T extends Number, EE extends Enum<? extends EE>> void test(Map<T, T> m1, Map<T, ? extends T> m2, Map<T, ? super T> m3, T t, EE enumValue) {
-//        a(m1);
-//        //a(m2); reason: Incompatible equality constraint: T and capture of ? extends T
-//        //a(m3); reason: Incompatible equality constraint: T and capture of ? super T
-//        b(m1);
-//        b(m2);
-//        //b(m3); // reason: no instance(s) of type variable(s) exist so that capture of ? super T conforms to Number
-//                 // inference variable N has incompatible bounds:
-//                 // equality constraints: T
-//        c(m1);
-//        //c(m2); // reason: no instance(s) of type variable(s) exist so that T conforms to capture of ? extends T
-//        c(m3);
-//        d(E1.A);
-//        //d(enumValue); // reason: no instance(s) of type variable(s) E exist so that ? extends EE can be converted to E
-//    }
-//}
-//
-//
-//
-//    class B<> {
-//
-//    }
-//}
-//
 
 @SuppressWarnings("ConstantConditions")
 class TypeUtilsTest implements RewriteTest {
@@ -1001,678 +906,5 @@ class TypeUtilsTest implements RewriteTest {
             }.visit(cu, new InMemoryExecutionContext()))
           )
         );
-    }
-
-    @Test
-    @SuppressWarnings("rawtypes")
-    void isOfType() {
-        rewriteRun(
-          java(
-            """
-              import java.util.List;
-              import java.util.Map;
-              
-              class Test<T extends Number, U extends List<String>, V extends U, X> {
-                  Integer integerA;
-                  Integer integerB;
-              
-                  int[] intArrayA;
-                  int[] intArrayB;
-                  Integer[] integerArrayA;
-                  Integer[] integerArrayB;
-                  String[] stringArrayA;
-                  String[] stringArrayB;
-                  List<String>[] genericArrayA;
-                  List<String>[] genericArrayB;
-                  Integer[][] nestedArrayA;
-                  Integer[][] nestedArrayB;
-                  T[] tArrayA;
-                  T[] tArrayB;
-                  U[] uArrayA;
-                  U[] uArrayB;
-                  V[] vArrayA;
-                  V[] vArrayB;
-                  X[] xArrayA;
-                  X[] xArrayB;
-              
-                  T numberTypeA;
-                  T numberTypeB;
-                  U listTypeA;
-                  U listTypeB;
-                  V nestedListTypeA;
-                  V nestedListTypeB;
-                  X genericA;
-                  X genericB;
-              
-                  List<T> numberListA;
-                  List<T> numberListB;
-                  List<String> listStringA;
-                  List<String> listStringB;
-                  Map<String, T> stringToNumberMapA;
-                  Map<String, T> stringToNumberMapB;
-                  Map<String, X> stringToGenericMapA;
-                  Map<String, X> stringToGenericMapB;
-              
-                  List<? extends Number> extendsNumberListA;
-                  List<? extends Number> extendsNumberListB;
-                  List<? super Integer> superIntegerListA;
-                  List<? super Integer> superIntegerListB;
-              
-                  Map<String, List<Map<Integer, String>>> complexNestedA;
-                  Map<String, List<Map<Integer, String>>> complexNestedB;
-              }
-              """,
-            spec -> spec.afterRecipe(cu -> new JavaIsoVisitor<>() {
-                @Override
-                public J.CompilationUnit visitCompilationUnit(J.CompilationUnit cu, Object o) {
-                    try (TypeUtilsAssertions assertions = new TypeUtilsAssertions(cu)) {
-                        // 1. Primitive exact matches
-                        assertions.isOfType("int", "int").isTrue();
-                        assertions.isOfType("int", "Integer").isFalse();
-                        assertions.isOfType("Integer", "int").isFalse();
-
-                        // 2. Array matches
-                        assertions.isOfType("int[]", "int[]").isTrue();
-                        assertions.isOfType("Integer[]", "Integer[]").isTrue();
-                        assertions.isOfType("Integer[]", "int[]").isFalse();
-                        assertions.isOfType("int[]", "Integer[]").isFalse();
-                        assertions.isOfType("Integer[][]", "Integer[][]").isTrue();
-                        assertions.isOfType("List<String>[]", "List<String>[]").isTrue();
-                        assertions.isOfType("List<String>[]", "String[]").isFalse();
-                        assertions.isOfType("int[]", "String[]").isFalse();
-                        assertions.isOfType("List<String>[]", "String[]").isFalse();
-
-                        // 3. Generic array matches
-                        assertions.isOfType("T[]", "T[]").isTrue();
-                        assertions.isOfType("U[]", "U[]").isTrue();
-                        assertions.isOfType("T[]", "Integer[]").isFalse();
-                        assertions.isOfType("U[]", "List<String>[]").isFalse();
-                        assertions.isOfType("Integer[][]", "T[]").isFalse();
-                        assertions.isOfType("T[]", "Integer[][]").isFalse();
-                        assertions.isOfType("U[]", "Integer[][]").isFalse();
-                        assertions.isOfType("U[]", "V[]").isFalse();
-                        assertions.isOfType("V[]", "U[]").isFalse();
-                        assertions.isOfType("Integer[][]", "int[]").isFalse();
-
-                        // 4. Type variable matches <T extends Number, U extends List<String>, V extends U>
-                        assertions.isOfType("T", "T").isTrue();
-                        assertions.isOfType("U", "U").isTrue();
-                        assertions.isOfType("V", "V").isTrue();
-                        assertions.isOfType("T", "Integer").isFalse();
-                        assertions.isOfType("T", "Integer").isFalse();
-                        assertions.isOfType("U", "V").isFalse();
-                        assertions.isOfType("T", "U").isFalse();
-
-                        // 5. Parameterized types
-                        assertions.isOfType("List<T>", "List<T>").isTrue();
-                        assertions.isOfType("List<? extends Number>", "List<? extends Number>").isTrue();
-                        assertions.isOfType("Map<String, List<Map<Integer, String>>>", "Map<String, List<Map<Integer, String>>>").isTrue();
-                        assertions.isOfType("List<T>", "List<? extends Number>").isFalse();
-                        assertions.isOfType("List<? extends Number>", "List<T>").isFalse();
-
-                        // 6. With INFER mode <T extends Number, U extends List<String>, V extends U>
-                        assertions.isOfType("T", "Integer", true).isTrue();
-                        assertions.isOfType("U", "Integer", true).isFalse();
-                        assertions.isOfType("U", "List<String>", true).isTrue();
-                        assertions.isOfType("V", "List<String>", true).isTrue();
-                        assertions.isOfType("T", "Integer[]", true).isFalse();
-                        assertions.isOfType("X", "Integer[]", true).isTrue();
-                        assertions.isOfType("T", "int[]", true).isFalse();
-                        assertions.isOfType("X", "int[]", true).isTrue();
-                        assertions.isOfType("T[]", "int[]", true).isFalse();
-                        assertions.isOfType("X[]", "int[]", true).isFalse();
-                        assertions.isOfType("T[]", "Integer[]", true).isTrue();
-                        assertions.isOfType("X[]", "Integer[]", true).isTrue();
-                        assertions.isOfType("U[]", "List<String>[]", true).isTrue();
-                        assertions.isOfType("V[]", "List<String>[]", true).isTrue();
-                        assertions.isOfType("Integer[][]", "T[]", true).isFalse();
-                        assertions.isOfType("X[]", "Integer[][]", true).isTrue();
-                        assertions.isOfType("T[]", "Integer[][]", true).isFalse();
-                        assertions.isOfType("U[]", "V[]", true).isTrue();
-                        assertions.isOfType("V[]", "U[]", true).isTrue();
-                        assertions.isOfType("Integer[][]", "int[]", true).isFalse();
-                        assertions.isOfType("Map<String, T>", "Map<String, List<Map<Integer, String>>>", true).isFalse();
-                        assertions.isOfType("Map<String, X>", "Map<String, List<Map<Integer, String>>>", true).isTrue();
-                        assertions.isOfType("Map<String, List<Map<Integer, String>>>", "Map<String, T>", true).isFalse();
-                    }
-                    return cu;
-                }
-            }.visit(cu, new InMemoryExecutionContext()))
-          )
-        );
-    }
-
-    @Test
-    void isClassAssignableTo() {
-        rewriteRun(
-          java(
-            """
-              import java.io.Serializable;
-              import java.util.ArrayList;
-              import java.util.Collection;
-              import java.util.List;
-              
-              @SuppressWarnings("all")
-              class Test<T extends Number & Serializable, U> {
-                  Integer integer;
-                  Boolean bool;
-                  Double bool;
-                  Number number;
-                  Cloneable cloneable;
-                  Serializable serializable;
-                  String[] array;
-              
-                  Object obj;
-                  String str;
-                  List listRaw;
-                  Collection collectionRaw;
-                  ArrayList arrayListRaw;
-                  List<String> listString;
-                  T genericBounded;
-                  U generic;
-              }
-              """,
-            spec -> spec.afterRecipe(cu -> new JavaIsoVisitor<>() {
-                @Override
-                public J.CompilationUnit visitCompilationUnit(J.CompilationUnit cu, Object o) {
-                    try (TypeUtilsAssertions assertions = new TypeUtilsAssertions(cu)) {
-                        // Boxed from primitives
-                        assertions.isAssignableTo("Integer", "int").isTrue();
-                        assertions.isAssignableTo("Number", "int").isTrue();
-                        assertions.isAssignableTo("Serializable", "int").isTrue();
-                        assertions.isAssignableTo("Boolean", "boolean").isTrue();
-                        assertions.isAssignableTo("Number", "boolean").isFalse();
-                        assertions.isAssignableTo("Serializable", "boolean").isTrue();
-                        assertions.isAssignableTo("Double", "double").isTrue();
-                        assertions.isAssignableTo("Number", "double").isTrue();
-                        assertions.isAssignableTo("Serializable", "double").isTrue();
-                        assertions.isAssignableTo("String", "int").isFalse();
-
-                        // FullyQualified direct
-                        assertions.isAssignableTo("Object", "String").isTrue();
-                        assertions.isAssignableTo("String", "Object").isFalse();
-                        assertions.isAssignableTo("List", "String").isFalse();
-
-                        // Null type (assignable to any reference type)
-                        assertions.isAssignableTo("String", "null").isTrue();
-                        assertions.isAssignableTo("List", "null").isTrue();
-
-                        // Parameterized type to raw type
-                        assertions.isAssignableTo("List", "List<String>").isTrue();
-
-                        // Class to interface
-                        assertions.isAssignableTo("Serializable", "String").isTrue();
-                        assertions.isAssignableTo("Collection", "ArrayList").isTrue();
-
-                        // Interface to class
-                        assertions.isAssignableTo("String", "Serializable").isFalse();
-
-                        // Array assignability
-                        assertions.isAssignableTo("Object", "String[]").isTrue();
-                        assertions.isAssignableTo("Cloneable", "String[]").isTrue();
-                        assertions.isAssignableTo("Serializable", "String[]").isTrue();
-
-                        // Generic type <T extends Number & Serializable, U>
-                        assertions.isAssignableTo("Serializable", "T").isTrue();
-                        assertions.isAssignableTo("Number", "T").isTrue();
-                        assertions.isAssignableTo("String", "T").isFalse();
-                        assertions.isAssignableTo("Object", "T").isTrue();
-                        assertions.isAssignableTo("Number", "U").isFalse();
-                        assertions.isAssignableTo("Number", "U").isFalse();
-                        assertions.isAssignableTo("Object", "U").isTrue();
-                    }
-                    return cu;
-                }
-            }.visit(cu, new InMemoryExecutionContext()))
-          )
-        );
-    }
-
-    @Test
-    @SuppressWarnings("rawtypes")
-    void isParameterizedAssignableTo() {
-        rewriteRun(
-          java(
-            """
-              import java.util.*;
-              import java.util.function.Supplier;
-              
-              class Test<T, U extends T, N extends Number, CS extends CharSequence> {
-                  ArrayList v1;
-                  Comparable<?> v2;
-                  Comparable<ImplementsComparable> v3;
-                  Comparable<Number> v4;
-                  Comparable<String> v5;
-                  ComparableSupplier<String, Number> v6;
-                  ExtendsComparable v7;
-                  List v8;
-                  List<? extends CharSequence> v9;
-                  List<? extends List<? extends CharSequence>> v10;
-                  List<? super String> v11;
-                  List<? super CharSequence> v25;
-                  List<? super T> v26;
-                  List<? super U> v27;
-                  List<?> v12;
-                  List<CS> v13;
-                  List<CharSequence> v14;
-                  List<List<? extends CharSequence>> v15;
-                  List<List<String>> v16;
-                  List<N> v17;
-                  List<String> v18;
-                  List<T> v19;
-                  List<U> v20;
-                  MySupplier<Number> v21;
-                  Supplier<Number> v22;
-                  Supplier<String> v23;
-                  ImplementsComparable v24;
-              
-                  static abstract class ImplementsComparable implements Comparable<ImplementsComparable> {}
-                  static abstract class ExtendsComparable extends ImplementsComparable {}
-                  static abstract class MySupplier<T> implements Supplier<T> {}
-                  static abstract class ComparableSupplier<T, U> extends MySupplier<U> implements Comparable<T> {}
-              }
-              """,
-            spec -> spec.afterRecipe(cu -> new JavaIsoVisitor<>() {
-                @Override
-                public J.CompilationUnit visitCompilationUnit(J.CompilationUnit cu, Object o) {
-                    try (TypeUtilsAssertions assertions = new TypeUtilsAssertions(cu)) {
-                        // 1. Generic Variance
-                        assertions.isAssignableTo("List<? extends CharSequence>", "List<String>").isTrue();
-                        assertions.isAssignableTo("List<String>", "List<? extends CharSequence>").isFalse();
-                        assertions.isAssignableTo("List<? super String>", "List<CharSequence>").isTrue();
-
-                        // 2. Wildcards and Raw Types
-                        assertions.isAssignableTo("List<?>", "List<String>").isTrue();
-                        assertions.isAssignableTo("List<?>", "ArrayList").isTrue();
-                        assertions.isAssignableTo("List<String>", "List").isFalse(); // We don't allow unsafe assignments
-                        assertions.isAssignableTo("List<?>", "List").isTrue(); // Except for wildcards
-
-                        // 3. Type Hierarchy with Generics (String, Number)
-                        assertions.isAssignableTo("Comparable<?>", "ImplementsComparable").isTrue();
-                        assertions.isAssignableTo("Comparable<ImplementsComparable>", "ImplementsComparable").isTrue();
-                        assertions.isAssignableTo("Comparable<ImplementsComparable>", "ExtendsComparable").isTrue();
-                        assertions.isAssignableTo("Comparable<?>", "ExtendsComparable").isTrue();
-                        assertions.isAssignableTo("Comparable<String>", "ExtendsComparable").isFalse();
-
-                        assertions.isAssignableTo("Comparable<String>", "ComparableSupplier<String, Number>").isTrue();
-                        assertions.isAssignableTo("Comparable<Number>", "ComparableSupplier<String, Number>").isFalse();
-                        assertions.isAssignableTo("Supplier<Number>", "ComparableSupplier<String, Number>").isTrue();
-                        assertions.isAssignableTo("Supplier<String>", "ComparableSupplier<String, Number>").isFalse();
-                        assertions.isAssignableTo("MySupplier<Number>", "ComparableSupplier<String, Number>").isTrue();
-                        assertions.isAssignableTo("Comparable<?>", "ComparableSupplier<String, Number>").isTrue();
-
-                        // 4. Type Variables
-                        assertions.isAssignableTo("List<T>", "List<String>").isFalse();
-                        assertions.isAssignableTo("List<T>", "List<U>").isFalse();
-                        assertions.isAssignableTo("List<? extends CharSequence>", "List<CS>").isTrue();
-                        assertions.isAssignableTo("List<? super U>", "List<? super T>").isTrue();
-                        assertions.isAssignableTo("List<? super String>", "List<? super CharSequence>").isTrue();
-                        assertions.isAssignableTo("List<? super T>", "List<? super U>").isFalse();
-                        assertions.isAssignableTo("List<? super CharSequence>", "List<? super String>").isFalse();
-
-                        // 5. Edge Cases
-                        assertions.isAssignableTo("List<? extends List<? extends CharSequence>>", "List<List<String>>").isTrue();
-                        assertions.isAssignableTo("List<List<? extends CharSequence>>", "List<List<String>>").isFalse();
-
-                        // 6. Inference Mode
-                        assertions.isAssignableTo("List<T>", "List<String>", true).isTrue();
-                        assertions.isAssignableTo("List<CS>", "List<String>", true).isTrue();
-                        assertions.isAssignableTo("List<N>", "List<String>", true).isFalse();
-                        assertions.isAssignableTo("List<? super T>", "List<? super String>", true).isTrue();
-                    }
-                    return cu;
-                }
-            }.visit(cu, new InMemoryExecutionContext()))
-          )
-        );
-    }
-
-    @Test
-    void isAssignableToArray() {
-        rewriteRun(
-          java(
-            """
-              class Test<T extends CharSequence, U, V extends Number> {
-                  Object[] objectArray;
-                  String[] stringArray;
-                  CharSequence[] charSequenceArray;
-                  int[] intArray;
-                  double[] doubleArray;
-                  Integer[] integerArray;
-                  Double[][] double2DArray;
-                  Number[][] number2DArray;
-                  Object[][] object2DArray;
-                  String[][] string2DArray;
-                  T[] genericCsArray;
-                  U[] genericArray;
-                  V[] genericNumericArray;
-                  U generic;
-              }
-              """,
-            spec -> spec.afterRecipe(cu -> new JavaIsoVisitor<>() {
-                @Override
-                public J.CompilationUnit visitCompilationUnit(J.CompilationUnit cu, Object o) {
-                    try (TypeUtilsAssertions assertions = new TypeUtilsAssertions(cu)) {
-                        // Identity and exact match
-                        assertions.isAssignableTo("String[]", "String[]").isTrue();
-
-                        // Covariant assignability of reference types
-                        assertions.isAssignableTo("Object[]", "String[]").isTrue();
-                        assertions.isAssignableTo("CharSequence[]", "String[]").isTrue();
-
-                        // Reverse should fail
-                        assertions.isAssignableTo("String[]", "Object[]").isFalse();
-                        assertions.isAssignableTo("String[]", "CharSequence[]").isFalse();
-
-                        // Primitive arrays are not assignable to Object[]
-                        assertions.isAssignableTo("Object[]", "int[]").isFalse();
-                        assertions.isAssignableTo("Object[]", "Integer[]").isTrue();
-
-                        // Primitive identity
-                        assertions.isAssignableTo("int[]", "int[]").isTrue();
-                        assertions.isAssignableTo("int[]", "Integer[]").isFalse();
-                        assertions.isAssignableTo("Integer[]", "int[]").isFalse();
-
-                        // Different primitives are not assignable
-                        assertions.isAssignableTo("int[]", "double[]").isFalse();
-
-                        // 2D array covariance
-                        assertions.isAssignableTo("Object[][]", "String[][]").isTrue();
-                        assertions.isAssignableTo("Number[][]", "Double[][]").isTrue();
-                        assertions.isAssignableTo("Double[][]", "Number[][]").isFalse();
-
-                        // Incompatible inner dimension
-                        assertions.isAssignableTo("Number[][]", "Integer[]").isFalse();
-
-                        // Generics: <T extends CharSequence, U>
-                        assertions.isAssignableTo("T[]", "String[]").isFalse();
-                        assertions.isAssignableTo("T[]", "CharSequence[]").isFalse();
-                        assertions.isAssignableTo("Object[]", "T[]").isTrue();
-                        assertions.isAssignableTo("CharSequence[]", "T[]").isTrue();
-                        assertions.isAssignableTo("U[]", "CharSequence[]").isFalse();
-                        assertions.isAssignableTo("Object[]", "U[]").isTrue();
-
-                        // Infer mode: <T extends CharSequence, U>
-                        assertions.isAssignableTo("T[]", "String[]", true).isTrue();
-                        assertions.isAssignableTo("T[]", "CharSequence[]", true).isTrue();
-                        assertions.isAssignableTo("T[]", "String[][]", true).isFalse();
-
-                        assertions.isAssignableTo("U", "String[]", true).isTrue();
-                        assertions.isAssignableTo("U", "CharSequence[]", true).isTrue();
-                        assertions.isAssignableTo("U", "String[][]", true).isTrue();
-                        assertions.isAssignableTo("U", "int[]", true).isTrue();
-                        assertions.isAssignableTo("U", "double[]", true).isTrue();
-
-                        assertions.isAssignableTo("U[]", "String[]", true).isTrue();
-                        assertions.isAssignableTo("U[]", "CharSequence[]", true).isTrue();
-                        assertions.isAssignableTo("U[]", "String[][]", true).isTrue();
-                        assertions.isAssignableTo("U[]", "int[]", true).isFalse();
-                        assertions.isAssignableTo("V[]", "int[]", true).isFalse();
-                        assertions.isAssignableTo("V[]", "Integer[]", true).isTrue();
-                        assertions.isAssignableTo("U[]", "double[]", true).isFalse();
-                    }
-                    return cu;
-                }
-            }.visit(cu, new InMemoryExecutionContext()))
-          )
-        );
-    }
-
-    @Test
-    void isAssignableToPrimitive() {
-        rewriteRun(
-          java(
-            """
-              class Test<T, U extends Number> {
-                  Byte boxedByte;
-                  Character boxedChar;
-                  Short boxedShort;
-                  Integer boxedInt;
-                  Long boxedLong;
-                  Float boxedFloat;
-                  Double boxedDouble;
-                  Boolean boxedBoolean;
-              
-                  T genericT;
-                  U genericU;
-              }
-              """,
-            spec -> spec.afterRecipe(cu -> new JavaIsoVisitor<>() {
-                @Override
-                public J.CompilationUnit visitCompilationUnit(J.CompilationUnit cu, Object o) {
-                    try (TypeUtilsAssertions assertions = new TypeUtilsAssertions(cu)) {
-                        // Direct primitive assignability
-                        assertions.isAssignableTo("int", "byte").isTrue();
-                        assertions.isAssignableTo("int", "char").isTrue();
-                        assertions.isAssignableTo("int", "short").isTrue();
-                        assertions.isAssignableTo("int", "int").isTrue();
-                        assertions.isAssignableTo("int", "long").isFalse();
-                        assertions.isAssignableTo("float", "int").isTrue();
-                        assertions.isAssignableTo("double", "float").isTrue();
-                        assertions.isAssignableTo("float", "double").isFalse();
-
-                        // Boolean isn't compatible with numeric types
-                        assertions.isAssignableTo("int", "boolean").isFalse();
-                        assertions.isAssignableTo("boolean", "boolean").isTrue();
-                        assertions.isAssignableTo("boolean", "int").isFalse();
-
-                        // Auto-unboxing
-                        assertions.isAssignableTo("int", "Integer").isTrue();
-                        assertions.isAssignableTo("double", "Double").isTrue();
-                        assertions.isAssignableTo("boolean", "Boolean").isTrue();
-                        assertions.isAssignableTo("Integer", "int").isTrue();
-                        assertions.isAssignableTo("Double", "double").isTrue();
-                        assertions.isAssignableTo("Boolean", "boolean").isTrue();
-
-                        // Mismatched boxed types
-                        assertions.isAssignableTo("int", "Boolean").isFalse();
-                        assertions.isAssignableTo("boolean", "Integer").isFalse();
-                        assertions.isAssignableTo("Boolean", "int").isFalse();
-                        assertions.isAssignableTo("Integer", "boolean").isFalse();
-
-                        // Generics <T, U extends Number>
-                        assertions.isAssignableTo("T", "byte", true).isTrue();
-                        assertions.isAssignableTo("T", "short", true).isTrue();
-                        assertions.isAssignableTo("T", "char", true).isTrue();
-                        assertions.isAssignableTo("T", "int", true).isTrue();
-                        assertions.isAssignableTo("T", "long", true).isTrue();
-                        assertions.isAssignableTo("T", "float", true).isTrue();
-                        assertions.isAssignableTo("T", "double", true).isTrue();
-                        assertions.isAssignableTo("T", "boolean", true).isTrue();
-
-                        assertions.isAssignableTo("U", "byte", true).isTrue();
-                        assertions.isAssignableTo("U", "short", true).isTrue();
-                        assertions.isAssignableTo("U", "int", true).isTrue();
-                        assertions.isAssignableTo("U", "long", true).isTrue();
-                        assertions.isAssignableTo("U", "float", true).isTrue();
-                        assertions.isAssignableTo("U", "double", true).isTrue();
-                        assertions.isAssignableTo("U", "char", true).isFalse();
-                        assertions.isAssignableTo("U", "boolean", true).isFalse();
-                    }
-                    return cu;
-                }
-            }.visit(cu, new InMemoryExecutionContext()))
-          )
-        );
-    }
-
-    @Test
-    void isAssignableToGenericTypeVariable() {
-        rewriteRun(
-          java(
-            """
-              class Test {
-                  class A<T, U extends T, V extends U, X> {
-                      T t;
-                      U u;
-                      V v;
-                      X x;
-                  }
-              
-                  class B<T, U extends T, V extends U, X> {
-                      T t;
-                      U u;
-                      V v;
-                      X x;
-                  }
-              }
-              """,
-            spec -> spec.afterRecipe(cu -> new JavaIsoVisitor<>() {
-                @Override
-                public J.CompilationUnit visitCompilationUnit(J.CompilationUnit cu, Object o) {
-                    try (TypeUtilsAssertions assertions = new TypeUtilsAssertions(cu)) {
-                        // 1. Same variable name
-                        assertions.isAssignableTo("T", "T").isTrue();
-                        assertions.isAssignableTo("U", "U").isTrue();
-                        assertions.isOfType("T", "T").isTrue();
-
-                        // 2. Different variables with compatible bounds
-                        // class <T, U extends T, V extends U>
-                        assertions.isAssignableTo("T", "U").isTrue(); // U is assignable to T
-                        assertions.isAssignableTo("U", "T").isFalse(); // T not assignable to U (U more specific)
-                        assertions.isAssignableTo("T", "V").isTrue(); // V -> U -> T
-                        assertions.isAssignableTo("U", "V").isTrue(); // V -> U
-                        assertions.isAssignableTo("V", "T").isFalse(); // T is more general
-
-                        // 3. Unrelated variables
-                        // class <T, X>
-                        assertions.isAssignableTo("T", "X").isFalse();
-                        assertions.isAssignableTo("X", "T").isFalse();
-
-                        // 4. isOfType tests for completeness
-                        assertions.isOfType("T", "T").isTrue();
-                        assertions.isOfType("U", "U").isTrue();
-                        assertions.isOfType("T", "U").isFalse();
-                        assertions.isOfType("U", "T").isFalse();
-                    }
-                    return cu;
-                }
-            }.visit(cu, new InMemoryExecutionContext()))
-          )
-        );
-    }
-
-    @Test
-    void isAssignableToRecursive() {
-        rewriteRun(
-          java(
-            """
-              abstract class Comp implements Comparable<Comp> {}
-              abstract class Ext extends Comp {}
-              enum EnumType { A, B, C }
-              
-              class Test<E extends Enum<E>, C extends Comparable<? super C>> {
-                  E e;
-                  C c;
-                  Comp comp;
-                  Ext ext;
-                  EnumType enumType;
-              }
-              """,
-            spec -> spec.afterRecipe(cu -> new JavaIsoVisitor<>() {
-                @Override
-                public J.CompilationUnit visitCompilationUnit(J.CompilationUnit cu, Object o) {
-                    try (TypeUtilsAssertions assertions = new TypeUtilsAssertions(cu)) {
-                        // 1. Same variable name
-                        assertions.isAssignableTo("E", "EnumType", false).isFalse();
-                        assertions.isAssignableTo("E", "EnumType", true).isTrue();
-
-                        assertions.isAssignableTo("C", "Comp", false).isFalse();
-                        assertions.isAssignableTo("C", "Ext", false).isFalse();
-
-                        assertions.isAssignableTo("C", "Comp", true).isTrue();
-                        assertions.isAssignableTo("C", "Ext", true).isTrue();
-                    }
-                    return cu;
-                }
-            }.visit(cu, new InMemoryExecutionContext()))
-          )
-        );
-    }
-
-    static class TypeUtilsAssertions implements AutoCloseable {
-        SoftAssertions softly = new SoftAssertions();
-        Map<String, List<JavaType>> types = new HashMap<>();
-
-        public TypeUtilsAssertions(J.CompilationUnit cu) {
-            EnumSet.complementOf(EnumSet.of(JavaType.Primitive.String, JavaType.Primitive.None))
-              .forEach(e -> types.put(e.getKeyword(), new ArrayList<>(Collections.singletonList(e))));
-            new JavaIsoVisitor<Integer>() {
-                @Override
-                public J.VariableDeclarations visitVariableDeclarations(J.VariableDeclarations multiVariable, Integer o) {
-                    if (multiVariable.getTypeExpression() != null) {
-                        String type = multiVariable.getTypeExpression().printTrimmed(getCursor());
-                        types.computeIfAbsent(type, k -> new ArrayList<>(2)).add(multiVariable.getTypeExpression().getType());
-                    }
-                    for (J.VariableDeclarations.NamedVariable variable : multiVariable.getVariables()) {
-                        types.computeIfAbsent(variable.getSimpleName(), k -> new ArrayList<>(2)).add(variable.getVariableType());
-                    }
-                    return super.visitVariableDeclarations(multiVariable, o);
-                }
-
-                @Override
-                public J.MethodInvocation visitMethodInvocation(J.MethodInvocation method, Integer o) {
-                    String type = method.printTrimmed(getCursor());
-                    types.computeIfAbsent(type, k -> new ArrayList<>(2)).add(method.getMethodType());
-                    return super.visitMethodInvocation(method, o);
-                }
-            }.visit(cu, 0);
-        }
-
-        public ObjectAssert<JavaType> type(String type) {
-            JavaType javaType = getFirst(type);
-            return softly.assertThat(javaType);
-        }
-
-        public BooleanAssert isAssignableTo(String to, String from) {
-            return isAssignableTo(to, from, false);
-        }
-
-        public BooleanAssert isAssignableTo(String to, String from, boolean capture) {
-            JavaType toType = getFirst(to);
-            JavaType fromType = getLast(from);
-            return softly.assertThat(new Types(capture).isAssignableTo(toType, fromType))
-              .describedAs("isAssignableTo(%s, %s, %s)", to, from, capture);
-        }
-
-        public BooleanAssert isOfType(String to, String from) {
-            return isOfType(to, from, false);
-        }
-
-        public BooleanAssert isOfType(String to, String from, boolean capture) {
-            JavaType toType = getFirst(to);
-            JavaType fromType = getLast(from);
-            return softly.assertThat(new Types(capture).isOfType(toType, fromType))
-              .describedAs("isOfType(%s, %s, %s)", to, from, capture);
-        }
-
-        public StringAssert toString(String type) {
-            JavaType javaType = getFirst(type);
-            return softly.assertThat(TypeUtils.toString(javaType))
-              .describedAs("toString(%s)", type);
-        }
-
-        public StringAssert toGenericTypeString(String type) {
-            JavaType javaType = getFirst(type);
-            return softly.assertThat(TypeUtils.toGenericTypeString((JavaType.GenericTypeVariable) javaType))
-              .describedAs("toGenericTypeString(%s)", type);
-        }
-
-        private JavaType getFirst(String type) {
-            return Optional.ofNullable(types.get(type))
-              .flatMap(list -> list.stream().findFirst())
-              .orElseThrow(() -> new IllegalArgumentException("Type not found: " + type));
-        }
-
-        private JavaType getLast(String type) {
-            return Optional.ofNullable(types.get(type))
-              .map(list -> list.get(list.size() - 1))
-              .orElseThrow(() -> new IllegalArgumentException("Type not found: " + type));
-        }
-
-        @Override
-        public void close() {
-            softly.assertAll();
-        }
     }
 }
